Unify the flag and flag count services
