+++ h/services/__init__.py
@@ -37,9 +37,6 @@ def includeme(config):
     )
     config.register_service_factory(".feature.feature_service_factory", name="feature")
     config.register_service_factory(".flag.flag_service_factory", name="flag")
-    config.register_service_factory(
-        ".flag_count.flag_count_service_factory", name="flag_count"
-    )
     config.register_service_factory(".group.groups_factory", name="group")
     config.register_service_factory(
         ".group_create.group_create_factory", name="group_create"
