+++ h/services/annotation_json_presentation/factory.py
@@ -11,6 +11,5 @@ def annotation_json_presentation_service_factory(_context, request):
         # Services
         links_svc=request.find_service(name="links"),
         flag_svc=request.find_service(name="flag"),
-        flag_count_svc=request.find_service(name="flag_count"),
         user_svc=request.find_service(name="user"),
     )
