+++ h/services/flag.py
@@ -1,9 +1,11 @@
-from h import models
+import sqlalchemy as sa
+
+from h.models import Flag
 
 
 class FlagService:
     def __init__(self, session):
-        self.session = session
+        self._session = session
 
     def flagged(self, user, annotation):
         """
@@ -18,9 +20,7 @@ class FlagService:
         :returns: True/False depending on the existence of a flag.
         :rtype: bool
         """
-        query = self.session.query(models.Flag).filter_by(
-            user=user, annotation=annotation
-        )
+        query = self._session.query(Flag).filter_by(user=user, annotation=annotation)
         return query.count() > 0
 
     def all_flagged(self, user, annotation_ids):
@@ -41,8 +41,8 @@ class FlagService:
         if not annotation_ids:
             return set()
 
-        query = self.session.query(models.Flag.annotation_id).filter(
-            models.Flag.annotation_id.in_(annotation_ids), models.Flag.user == user
+        query = self._session.query(Flag.annotation_id).filter(
+            Flag.annotation_id.in_(annotation_ids), Flag.user == user
         )
 
         return {f.annotation_id for f in query}
@@ -65,8 +65,50 @@ class FlagService:
         if self.flagged(user, annotation):
             return
 
-        flag = models.Flag(user=user, annotation=annotation)
-        self.session.add(flag)
+        flag = Flag(user=user, annotation=annotation)
+        self._session.add(flag)
+
+    def flag_count(self, annotation):
+        """
+        Return the number of times a given annotation has been flagged.
+
+        :param annotation: The annotation to check for flags.
+        :type annotation: h.models.Annotation
+
+        :returns: The number of times the annotation has been flagged.
+        :rtype: int
+        """
+        return (
+            self._session.query(sa.func.count(Flag.id))
+            .filter(Flag.annotation_id == annotation.id)
+            .scalar()
+        )
+
+    def flag_counts(self, annotation_ids):
+        """
+        Return flag counts for a batch of annotations.
+
+        :param annotation_ids: The IDs of the annotations to check.
+        :type annotation_ids: sequence of unicode
+
+        :returns: A map of annotation IDs to flag counts.
+        :rtype: dict[unicode, int]
+        """
+        if not annotation_ids:
+            return {}
+
+        query = (
+            self._session.query(
+                sa.func.count(Flag.id).label("flag_count"), Flag.annotation_id
+            )
+            .filter(Flag.annotation_id.in_(annotation_ids))
+            .group_by(Flag.annotation_id)
+        )
+
+        flag_counts = {f.annotation_id: f.flag_count for f in query}
+        missing_ids = set(annotation_ids) - set(flag_counts.keys())
+        flag_counts.update({id_: 0 for id_ in missing_ids})
+        return flag_counts
 
 
 def flag_service_factory(_context, request):
